+++ packages/example/src/index.test.ts
@@ -1,7 +1,0 @@
-import { expect } from 'earl'
-
-describe('plus operator', () => {
-  it('adds two numbers', () => {
-    expect(1 + 2).toEqual(3)
-  })
-})